refactor the ARXML code
This commit basically comes down to a lot of copy and paste: The
individual ARXML classes previously contained in `arxml.py` are
separated into different files and the submodule thus gets a dedicated
directory.
