+++ cantools/database/can/formats/arxml/bus_specifics.py
@@ -1,0 +1,6 @@
+class AutosarBusSpecifics(object):
+    """This class collects the AUTOSAR specific information of a CAN bus
+
+    """
+    def __init__(self):
+        pass
