+++ cantools/database/can/formats/arxml/ecu_extract_loader.py
@@ -1,0 +1,372 @@
+# Load an ECU extract CAN database from an ARXML formatted file.
+import logging
+
+from typing import Any, List
+from xml.etree import ElementTree
+
+from ....utils import type_sort_signals, sort_signals_by_start_bit
+from ...bus import Bus
+from ...message import Message
+from ...signal import Signal
+from ...internal_database import InternalDatabase
+
+from decimal import Decimal
+from ...signal import Decimal as SignalDecimal
+
+def make_xpath(location: List[str]) -> str:
+    """Convenience function to traverse the XML element tree more easily
+
+    (This function is only used by the EcuExtractLoader.)"""
+    return './ns:' + '/ns:'.join(location)
+
+LOGGER = logging.getLogger(__name__)
+
+# The ARXML XML namespace for the EcuExtractLoader
+NAMESPACE = 'http://autosar.org/schema/r4.0'
+NAMESPACES = {'ns': NAMESPACE}
+
+ECUC_VALUE_COLLECTION_XPATH = make_xpath([
+    'AR-PACKAGES',
+    'AR-PACKAGE',
+    'ELEMENTS',
+    'ECUC-VALUE-COLLECTION'
+])
+ECUC_MODULE_CONFIGURATION_VALUES_REF_XPATH = make_xpath([
+    'ECUC-VALUES',
+    'ECUC-MODULE-CONFIGURATION-VALUES-REF-CONDITIONAL',
+    'ECUC-MODULE-CONFIGURATION-VALUES-REF'
+])
+ECUC_REFERENCE_VALUE_XPATH = make_xpath([
+    'REFERENCE-VALUES',
+    'ECUC-REFERENCE-VALUE'
+])
+DEFINITION_REF_XPATH = make_xpath(['DEFINITION-REF'])
+VALUE_XPATH = make_xpath(['VALUE'])
+VALUE_REF_XPATH = make_xpath(['VALUE-REF'])
+SHORT_NAME_XPATH = make_xpath(['SHORT-NAME'])
+PARAMETER_VALUES_XPATH = make_xpath(['PARAMETER-VALUES'])
+REFERENCE_VALUES_XPATH = make_xpath([
+    'REFERENCE-VALUES'
+])
+
+class EcuExtractLoader(object):
+
+    def __init__(self,
+                 root:Any,
+                 strict:bool,
+                 sort_signals:type_sort_signals=sort_signals_by_start_bit):
+        self.root = root
+        self.strict = strict
+        self.sort_signals = sort_signals
+
+    def load(self) -> InternalDatabase:
+        buses:List[Bus] = []
+        messages = []
+        version = None
+
+        ecuc_value_collection = self.root.find(ECUC_VALUE_COLLECTION_XPATH,
+                                               NAMESPACES)
+        values_refs = ecuc_value_collection.iterfind(
+            ECUC_MODULE_CONFIGURATION_VALUES_REF_XPATH,
+            NAMESPACES)
+        com_xpaths = [
+            value_ref.text
+            for value_ref in values_refs
+            if value_ref.text.endswith('/Com')
+        ]
+
+        if len(com_xpaths) != 1:
+            raise ValueError(
+                'Expected 1 /Com, but got {}.'.format(len(com_xpaths)))
+
+        com_config = self.find_com_config(com_xpaths[0] + '/ComConfig')
+
+        for ecuc_container_value in com_config:
+            definition_ref = ecuc_container_value.find(DEFINITION_REF_XPATH,
+                                                       NAMESPACES).text
+
+            if not definition_ref.endswith('ComIPdu'):
+                continue
+
+            message = self.load_message(ecuc_container_value)
+
+            if message is not None:
+                messages.append(message)
+
+        return InternalDatabase(messages,
+                                [],
+                                buses,
+                                version)
+
+    def load_message(self, com_i_pdu):
+        # Default values.
+        interval = None
+        senders = []
+        comments = None
+
+        # Name, frame id, length and is_extended_frame.
+        name = com_i_pdu.find(SHORT_NAME_XPATH, NAMESPACES).text
+        direction = None
+
+        for parameter, value in self.iter_parameter_values(com_i_pdu):
+            if parameter == 'ComIPduDirection':
+                direction = value
+                break
+
+        com_pdu_id_ref = None
+
+        for reference, value in self.iter_reference_values(com_i_pdu):
+            if reference == 'ComPduIdRef':
+                com_pdu_id_ref = value
+                break
+
+        if com_pdu_id_ref is None:
+            raise ValueError('No ComPduIdRef reference found.')
+
+        if direction == 'SEND':
+            frame_id, length, is_extended_frame = self.load_message_tx(
+                com_pdu_id_ref)
+        elif direction == 'RECEIVE':
+            frame_id, length, is_extended_frame = self.load_message_rx(
+                com_pdu_id_ref)
+        else:
+            raise NotImplementedError(
+                'Direction {} not supported.'.format(direction))
+
+        if frame_id is None:
+            LOGGER.warning('No frame id found for message %s.', name)
+
+            return None
+
+        if is_extended_frame is None:
+            LOGGER.warning('No frame type found for message %s.', name)
+
+            return None
+
+        if length is None:
+            LOGGER.warning('No length found for message %s.', name)
+
+            return None
+
+        # ToDo: interval, senders, comments
+
+        # Find all signals in this message.
+        signals = []
+        values = com_i_pdu.iterfind(ECUC_REFERENCE_VALUE_XPATH,
+                                    NAMESPACES)
+
+        for value in values:
+            definition_ref = value.find(DEFINITION_REF_XPATH,
+                                        NAMESPACES).text
+            if not definition_ref.endswith('ComIPduSignalRef'):
+                continue
+
+            value_ref = value.find(VALUE_REF_XPATH, NAMESPACES)
+            signal = self.load_signal(value_ref.text)
+
+            if signal is not None:
+                signals.append(signal)
+
+        return Message(frame_id=frame_id,
+                       is_extended_frame=is_extended_frame,
+                       name=name,
+                       length=length,
+                       senders=senders,
+                       send_type=None,
+                       cycle_time=interval,
+                       signals=signals,
+                       comment=comments,
+                       bus_name=None,
+                       strict=self.strict,
+                       sort_signals=self.sort_signals)
+
+    def load_message_tx(self, com_pdu_id_ref):
+        return self.load_message_rx_tx(com_pdu_id_ref,
+                                       'CanIfTxPduCanId',
+                                       'CanIfTxPduDlc',
+                                       'CanIfTxPduCanIdType')
+
+    def load_message_rx(self, com_pdu_id_ref):
+        return self.load_message_rx_tx(com_pdu_id_ref,
+                                       'CanIfRxPduCanId',
+                                       'CanIfRxPduDlc',
+                                       'CanIfRxPduCanIdType')
+
+    def load_message_rx_tx(self,
+                           com_pdu_id_ref,
+                           parameter_can_id,
+                           parameter_dlc,
+                           parameter_can_id_type):
+        can_if_tx_pdu_cfg = self.find_can_if_rx_tx_pdu_cfg(com_pdu_id_ref)
+        frame_id = None
+        length = None
+        is_extended_frame = None
+
+        if can_if_tx_pdu_cfg is not None:
+            for parameter, value in self.iter_parameter_values(can_if_tx_pdu_cfg):
+                if parameter == parameter_can_id:
+                    frame_id = int(value)
+                elif parameter == parameter_dlc:
+                    length = int(value)
+                elif parameter == parameter_can_id_type:
+                    is_extended_frame = (value == 'EXTENDED_CAN')
+
+        return frame_id, length, is_extended_frame
+
+    def load_signal(self, xpath):
+        ecuc_container_value = self.find_value(xpath)
+        if ecuc_container_value is None:
+            return None
+
+        name = ecuc_container_value.find(SHORT_NAME_XPATH, NAMESPACES).text
+
+        # Default values.
+        is_signed = False
+        is_float = False
+        minimum = None
+        maximum = None
+        factor = 1.0
+        offset = 0.0
+        unit = None
+        choices = None
+        comments = None
+        receivers = []
+        decimal = SignalDecimal(Decimal(factor), Decimal(offset))
+
+        # Bit position, length, byte order, is_signed and is_float.
+        bit_position = None
+        length = None
+        byte_order = None
+
+        for parameter, value in self.iter_parameter_values(ecuc_container_value):
+            if parameter == 'ComBitPosition':
+                bit_position = int(value)
+            elif parameter == 'ComBitSize':
+                length = int(value)
+            elif parameter == 'ComSignalEndianness':
+                byte_order = value.lower()
+            elif parameter == 'ComSignalType':
+                if value in ['SINT8', 'SINT16', 'SINT32']:
+                    is_signed = True
+                elif value in ['FLOAT32', 'FLOAT64']:
+                    is_float = True
+
+        if bit_position is None:
+            LOGGER.warning('No bit position found for signal %s.',name)
+
+            return None
+
+        if length is None:
+            LOGGER.warning('No bit size found for signal %s.', name)
+
+            return None
+
+        if byte_order is None:
+            LOGGER.warning('No endianness found for signal %s.', name)
+
+            return None
+
+        # ToDo: minimum, maximum, factor, offset, unit, choices,
+        #       comments and receivers.
+
+        return Signal(name=name,
+                      start=bit_position,
+                      length=length,
+                      receivers=receivers,
+                      byte_order=byte_order,
+                      is_signed=is_signed,
+                      scale=factor,
+                      offset=offset,
+                      minimum=minimum,
+                      maximum=maximum,
+                      unit=unit,
+                      choices=choices,
+                      comment=comments,
+                      is_float=is_float,
+                      decimal=decimal)
+
+    def find_com_config(self, xpath):
+        return self.root.find(make_xpath([
+            "AR-PACKAGES",
+            "AR-PACKAGE/[ns:SHORT-NAME='{}']".format(xpath.split('/')[1]),
+            "ELEMENTS",
+            "ECUC-MODULE-CONFIGURATION-VALUES/[ns:SHORT-NAME='Com']",
+            "CONTAINERS",
+            "ECUC-CONTAINER-VALUE/[ns:SHORT-NAME='ComConfig']",
+            "SUB-CONTAINERS"
+        ]),
+                              NAMESPACES)
+
+    def find_value(self, xpath):
+        return self.root.find(make_xpath([
+            "AR-PACKAGES",
+            "AR-PACKAGE/[ns:SHORT-NAME='{}']".format(xpath.split('/')[1]),
+            "ELEMENTS",
+            "ECUC-MODULE-CONFIGURATION-VALUES/[ns:SHORT-NAME='Com']",
+            "CONTAINERS",
+            "ECUC-CONTAINER-VALUE/[ns:SHORT-NAME='ComConfig']",
+            "SUB-CONTAINERS",
+            "ECUC-CONTAINER-VALUE/[ns:SHORT-NAME='{}']".format(xpath.split('/')[-1])
+        ]),
+                              NAMESPACES)
+
+    def find_can_if_rx_tx_pdu_cfg(self, com_pdu_id_ref):
+        messages = self.root.iterfind(
+            make_xpath([
+                "AR-PACKAGES",
+                "AR-PACKAGE/[ns:SHORT-NAME='{}']".format(
+                    com_pdu_id_ref.split('/')[1]),
+                "ELEMENTS",
+                "ECUC-MODULE-CONFIGURATION-VALUES/[ns:SHORT-NAME='CanIf']",
+                'CONTAINERS',
+                "ECUC-CONTAINER-VALUE/[ns:SHORT-NAME='CanIfInitCfg']",
+                'SUB-CONTAINERS',
+                'ECUC-CONTAINER-VALUE'
+            ]),
+            NAMESPACES)
+
+        for message in messages:
+            definition_ref = message.find(DEFINITION_REF_XPATH,
+                                          NAMESPACES).text
+
+            if definition_ref.endswith('CanIfTxPduCfg'):
+                expected_reference = 'CanIfTxPduRef'
+            elif definition_ref.endswith('CanIfRxPduCfg'):
+                expected_reference = 'CanIfRxPduRef'
+            else:
+                continue
+
+            for reference, value in self.iter_reference_values(message):
+                if reference == expected_reference:
+                    if value == com_pdu_id_ref:
+                        return message
+
+    def iter_parameter_values(self, param_conf_container):
+        parameters = param_conf_container.find(PARAMETER_VALUES_XPATH,
+                                               NAMESPACES)
+
+        if parameters is None:
+            raise ValueError('PARAMETER-VALUES does not exist.')
+
+        for parameter in parameters:
+            definition_ref = parameter.find(DEFINITION_REF_XPATH,
+                                            NAMESPACES).text
+            value = parameter.find(VALUE_XPATH, NAMESPACES).text
+            name = definition_ref.split('/')[-1]
+
+            yield name, value
+
+    def iter_reference_values(self, param_conf_container):
+        references = param_conf_container.find(REFERENCE_VALUES_XPATH,
+                                               NAMESPACES)
+
+        if references is None:
+            raise ValueError('REFERENCE-VALUES does not exist.')
+
+        for reference in references:
+            definition_ref = reference.find(DEFINITION_REF_XPATH,
+                                            NAMESPACES).text
+            value = reference.find(VALUE_REF_XPATH, NAMESPACES).text
+            name = definition_ref.split('/')[-1]
+
+            yield name, value
