+++ cantools/database/can/formats/arxml/system_loader.py
@@ -1,302 +1,31 @@
-# Load and dump a CAN database in ARXML format.
-
+# Load a CAN database in ARXML format.
 import re
 import logging
 import numbers
 from decimal import Decimal
-from typing import Union, Any, List, Optional
+from typing import Any, List, Optional
 from copy import deepcopy
 
 from xml.etree import ElementTree
 
-from ..signal import Signal, NamedSignalValue
-from ..signal import Decimal as SignalDecimal
-from ..message import Message
-from ..node import Node
-from ..bus import Bus
-from ..internal_database import InternalDatabase
-from ...utils import type_sort_signals, sort_signals_by_start_bit
+from .utils import parse_number_string
+from .database_specifics import AutosarDatabaseSpecifics
+from .bus_specifics import AutosarBusSpecifics
+from .node_specifics import AutosarNodeSpecifics
+from .message_specifics import AutosarMessageSpecifics
+
+from .secoc_properties import AutosarSecOCProperties
+from .end_to_end_properties import AutosarEnd2EndProperties
+
+from ...signal import Signal, NamedSignalValue
+from ...signal import Decimal as SignalDecimal
+from ...message import Message
+from ...node import Node
+from ...bus import Bus
+from ...internal_database import InternalDatabase
+from ....utils import type_sort_signals, sort_signals_by_start_bit
 
 LOGGER = logging.getLogger(__name__)
-
-class AutosarDatabaseSpecifics(object):
-    """This class collects the AUTOSAR specific information of a system
-
-    """
-    def __init__(self,
-                 arxml_version):
-        self._arxml_version = arxml_version
-
-    @property
-    def arxml_version(self):
-        """The used version of ARXML file format
-
-        Note that due to technical reasons we always return version
-        "4.0.0" for AUTOSAR 4.X.
-        """
-        return self._arxml_version
-
-class AutosarBusSpecifics(object):
-    """This class collects the AUTOSAR specific information of a CAN bus
-
-    """
-    def __init__(self):
-        pass
-
-class AutosarNodeSpecifics(object):
-    """This class collects the AUTOSAR specific information of node that
-    is attached to a CAN bus
-
-    AUTOSAR calls such nodes "ECU instances"...
-    """
-    def __init__(self):
-        pass
-
-class AutosarMessageSpecifics(object):
-    """This class collects all AUTOSAR specific information of a CAN message
-
-    This means useful information about CAN messages which is provided
-    by ARXML files, but is specific to AUTOSAR.
-    """
-
-    def __init__(self):
-        self._pdu_paths = []
-        self._is_nm = False
-        self._is_general_purpose = False
-        self._secoc: Optional[AutosarSecOCProperties] = None
-        self._e2e: Optional[AutosarEnd2EndProperties] = None
-        self._signal_group = None
-
-    @property
-    def pdu_paths(self):
-        """The ARXML paths of all PDUs featured by this message.
-
-        For the vast majority of messages, this list only has a single
-        entry. Messages with multiplexers and container frames are
-        different, though.
-        """
-        return self._pdu_paths
-
-    @property
-    def is_nm(self):
-        """True iff the message is used for network management
-        """
-        return self._is_nm
-
-    @property
-    def is_general_purpose(self):
-        """True iff the message is not used for signal-based communication
-
-        This comprises messages used for diagnostic and calibration
-        purpuses, e.g. messages used for the ISO-TP or XCP protocols.
-
-        """
-        return self._is_general_purpose
-
-    @property
-    def is_secured(self):
-        """True iff the message integrity is secured using SecOC
-        """
-        return self._secoc is not None
-
-    @property
-    def secoc(self):
-        """The properties required to implement secured on-board communication
-        """
-        return self._secoc
-
-    @property
-    def e2e(self) -> Optional['AutosarEnd2EndProperties']:
-        """Returns the end-to-end protection properties for the message"""
-        return self._e2e
-
-    @e2e.setter
-    def e2e(self, value: Optional['AutosarEnd2EndProperties']) -> None:
-        self._e2e = value
-
-class AutosarSecOCProperties:
-    """This class collects all attributes that are required to implement the
-    AUTOSAR secure on-board communication (SecOC) specification.
-
-    Be aware that the AUTOSAR SecOC specification does not cover the
-    required cryptographic primitives themselves, just the
-    "scaffolding" around them...
-    """
-    def __init__(self,
-                 auth_algorithm_name: Optional[str],
-                 freshness_algorithm_name: Optional[str],
-                 payload_length: Optional[int],
-                 data_id: Optional[int],
-                 auth_tx_bit_length: Optional[int],
-                 freshness_bit_length: Optional[int],
-                 freshness_tx_bit_length: Optional[int],
-                 ):
-
-        self._auth_algorithm_name = auth_algorithm_name
-        self._freshness_algorithm_name = freshness_algorithm_name
-
-        self._payload_length = payload_length
-        self._data_id = data_id
-
-        self._freshness_bit_length = freshness_bit_length
-        self._freshness_tx_bit_length = freshness_tx_bit_length
-        self._auth_tx_bit_length = auth_tx_bit_length
-
-    @property
-    def freshness_algorithm_name(self) -> Optional[str]:
-        """The name of the algorithm used for verifying the freshness of a
-        message.
-
-        This can be used to prevent replay attacks. Note that the
-        algorithms themselves are manufacturer-specific, i.e., AUTOSAR
-        does not define *any* freshness schemes.
-        """
-        return self._freshness_algorithm_name
-
-    @property
-    def auth_algorithm_name(self) -> Optional[str]:
-        """The name of the algorithm used for authentication
-
-        Note that the algorithms themselves are manufacturer-specific,
-        i.e., AUTOSAR does not define *any* cryptographic schemes.
-        """
-        return self._auth_algorithm_name
-
-    @property
-    def payload_length(self) -> Optional[int]:
-        """Returns the number of bytes covered by the payload of the secured
-        message
-
-        (The full message length is the length of the payload plus the
-        size of the security trailer.)
-        """
-        return self._payload_length
-
-    @property
-    def data_id(self) -> Optional[int]:
-        """The data ID required for authentication.
-
-        Be aware that this is a different data ID than that required
-        for End-To-End protection.
-        """
-        return self._data_id
-
-    @property
-    def freshness_bit_length(self) -> Optional[int]:
-        """The number of bits of the full freshness counter.
-        """
-        return self._freshness_bit_length
-
-    @property
-    def freshness_tx_bit_length(self) -> Optional[int]:
-        """The number of least-significant bits of the freshness counter that
-        is send as part of the secured frame.
-
-        This number is at most as large as the number of bits of
-        freshness counter objects.
-
-        """
-        return self._freshness_tx_bit_length
-
-    @property
-    def auth_tx_bit_length(self) -> Optional[int]:
-        """The number of most significant bits of the authenticator object
-        send as part of the secured frame
-
-        This is at most the length of the authenicator.
-        """
-        return self._auth_tx_bit_length
-
-
-class AutosarEnd2EndProperties:
-    """This class collects all attributes that are required to implement
-    AUTOSAR-conformant End-to-End protection (CRCs) of messages
-    """
-
-    def __init__(self):
-        self._category: Optional[str] = None
-        self._data_ids: Optional[List[int]] = None
-        self._payload_length: int = 0
-
-    @property
-    def category(self) -> Optional[str]:
-        """The category string of the applicable end-to-end protection
-        mechanism
-
-        Note that the contents of these are not specified by the
-        AUTOSAR standard.
-        """
-        return self._category
-
-    @category.setter
-    def category(self, value: Optional[str]) -> None:
-        self._category = value
-
-    @property
-    def data_ids(self) -> Optional[List[int]]:
-        """The list of data IDs applicable
-        """
-        return self._data_ids
-
-    @data_ids.setter
-    def data_ids(self, value: Optional[List[int]]) -> None:
-        self._data_ids = value
-
-    @property
-    def payload_length(self) -> int:
-        """The size of the end-to-end protected data in bytes
-
-        This number includes the end-to-end protection signals
-        themselves (i.e. the sequence counter and the CRC value)
-
-        """
-        return self._payload_length
-
-    @payload_length.setter
-    def payload_length(self, value: int) -> None:
-        self._payload_length = value
-
-def parse_number_string(in_string : str, allow_float : bool=False) \
-    -> Union[int, float]:
-    """Convert a string representing numeric value that is specified
-    within an ARXML file to either an integer or a floating point object
-
-    This is surprisingly complicated:
-
-    - Some ARXML files use "true" and "false" synonymous to 1 and 0
-    - ARXML uses the C notation (leading 0) to specify octal numbers
-      whereas python only accepts the "0o" prefix
-    - Some ARXML editors seem to sometimes include a dot in integer
-      numbers (e.g., they produce "123.0" instead of "123")
-    """
-
-    # the string literals "true" and "false" are interpreted as 1 and 0
-    if in_string == 'true':
-        return 1
-    elif in_string == 'false':
-        return 0
-    # the input string contains a dot -> floating point value
-    elif '.' in in_string:
-        tmp = float(in_string)
-
-        if not allow_float and tmp != int(tmp):
-            raise ValueError('Floating point value specified where integer '
-                             'is required')
-
-        # if an integer is required but a .0 floating point value is
-        # specified, we accept the input anyway. (this seems to be an
-        # ambiguity in the AUTOSAR specification.)
-        return tmp if allow_float else int(tmp)
-
-    in_string = in_string.strip()
-    if not in_string:
-        return 0
-    elif in_string[0] == '0' and in_string[1:2].isdigit():
-        # interpret strings starting with a 0 as octal because
-        # python's int(*, 0) does not for some reason.
-        return int(in_string, 8)
-
-    return int(in_string, 0) # autodetect the base
 
 class SystemLoader(object):
     def __init__(self,
@@ -1719,6 +1448,7 @@ class SystemLoader(object):
                                                 i_signal_spec)
         # Name, start position, length and byte order.
         name = self._load_signal_name(i_signal)
+
         start_position = \
             self._load_signal_start_position(i_signal_to_i_pdu_mapping)
         length = self._load_signal_length(i_signal, system_signal)
@@ -2601,417 +2331,3 @@ class SystemLoader(object):
                                                '&SW-DATA-DEF-PROPS-CONDITIONAL',
                                                '&BASE-TYPE'
                                             ])
-
-# The ARXML XML namespace for the EcuExtractLoader
-NAMESPACE = 'http://autosar.org/schema/r4.0'
-NAMESPACES = {'ns': NAMESPACE}
-
-ROOT_TAG = '{{{}}}AUTOSAR'.format(NAMESPACE)
-
-# ARXML XPATHs used by the EcuExtractLoader
-def make_xpath(location):
-    return './ns:' + '/ns:'.join(location)
-
-ECUC_VALUE_COLLECTION_XPATH = make_xpath([
-    'AR-PACKAGES',
-    'AR-PACKAGE',
-    'ELEMENTS',
-    'ECUC-VALUE-COLLECTION'
-])
-ECUC_MODULE_CONFIGURATION_VALUES_REF_XPATH = make_xpath([
-    'ECUC-VALUES',
-    'ECUC-MODULE-CONFIGURATION-VALUES-REF-CONDITIONAL',
-    'ECUC-MODULE-CONFIGURATION-VALUES-REF'
-])
-ECUC_REFERENCE_VALUE_XPATH = make_xpath([
-    'REFERENCE-VALUES',
-    'ECUC-REFERENCE-VALUE'
-])
-DEFINITION_REF_XPATH = make_xpath(['DEFINITION-REF'])
-VALUE_XPATH = make_xpath(['VALUE'])
-VALUE_REF_XPATH = make_xpath(['VALUE-REF'])
-SHORT_NAME_XPATH = make_xpath(['SHORT-NAME'])
-PARAMETER_VALUES_XPATH = make_xpath(['PARAMETER-VALUES'])
-REFERENCE_VALUES_XPATH = make_xpath([
-    'REFERENCE-VALUES'
-])
-
-class EcuExtractLoader(object):
-
-    def __init__(self,
-                 root:Any,
-                 strict:bool,
-                 sort_signals:type_sort_signals=sort_signals_by_start_bit):
-        self.root = root
-        self.strict = strict
-        self.sort_signals = sort_signals
-
-    def load(self) -> InternalDatabase:
-        buses:List[Bus] = []
-        messages = []
-        version = None
-
-        ecuc_value_collection = self.root.find(ECUC_VALUE_COLLECTION_XPATH,
-                                               NAMESPACES)
-        values_refs = ecuc_value_collection.iterfind(
-            ECUC_MODULE_CONFIGURATION_VALUES_REF_XPATH,
-            NAMESPACES)
-        com_xpaths = [
-            value_ref.text
-            for value_ref in values_refs
-            if value_ref.text.endswith('/Com')
-        ]
-
-        if len(com_xpaths) != 1:
-            raise ValueError(
-                'Expected 1 /Com, but got {}.'.format(len(com_xpaths)))
-
-        com_config = self.find_com_config(com_xpaths[0] + '/ComConfig')
-
-        for ecuc_container_value in com_config:
-            definition_ref = ecuc_container_value.find(DEFINITION_REF_XPATH,
-                                                       NAMESPACES).text
-
-            if not definition_ref.endswith('ComIPdu'):
-                continue
-
-            message = self.load_message(ecuc_container_value)
-
-            if message is not None:
-                messages.append(message)
-
-        return InternalDatabase(messages,
-                                [],
-                                buses,
-                                version)
-
-    def load_message(self, com_i_pdu):
-        # Default values.
-        interval = None
-        senders = []
-        comments = None
-
-        # Name, frame id, length and is_extended_frame.
-        name = com_i_pdu.find(SHORT_NAME_XPATH, NAMESPACES).text
-        direction = None
-
-        for parameter, value in self.iter_parameter_values(com_i_pdu):
-            if parameter == 'ComIPduDirection':
-                direction = value
-                break
-
-        com_pdu_id_ref = None
-
-        for reference, value in self.iter_reference_values(com_i_pdu):
-            if reference == 'ComPduIdRef':
-                com_pdu_id_ref = value
-                break
-
-        if com_pdu_id_ref is None:
-            raise ValueError('No ComPduIdRef reference found.')
-
-        if direction == 'SEND':
-            frame_id, length, is_extended_frame = self.load_message_tx(
-                com_pdu_id_ref)
-        elif direction == 'RECEIVE':
-            frame_id, length, is_extended_frame = self.load_message_rx(
-                com_pdu_id_ref)
-        else:
-            raise NotImplementedError(
-                'Direction {} not supported.'.format(direction))
-
-        if frame_id is None:
-            LOGGER.warning('No frame id found for message %s.', name)
-
-            return None
-
-        if is_extended_frame is None:
-            LOGGER.warning('No frame type found for message %s.', name)
-
-            return None
-
-        if length is None:
-            LOGGER.warning('No length found for message %s.', name)
-
-            return None
-
-        # ToDo: interval, senders, comments
-
-        # Find all signals in this message.
-        signals = []
-        values = com_i_pdu.iterfind(ECUC_REFERENCE_VALUE_XPATH,
-                                    NAMESPACES)
-
-        for value in values:
-            definition_ref = value.find(DEFINITION_REF_XPATH,
-                                        NAMESPACES).text
-            if not definition_ref.endswith('ComIPduSignalRef'):
-                continue
-
-            value_ref = value.find(VALUE_REF_XPATH, NAMESPACES)
-            signal = self.load_signal(value_ref.text)
-
-            if signal is not None:
-                signals.append(signal)
-
-        return Message(frame_id=frame_id,
-                       is_extended_frame=is_extended_frame,
-                       name=name,
-                       length=length,
-                       senders=senders,
-                       send_type=None,
-                       cycle_time=interval,
-                       signals=signals,
-                       comment=comments,
-                       bus_name=None,
-                       strict=self.strict,
-                       sort_signals=self.sort_signals)
-
-    def load_message_tx(self, com_pdu_id_ref):
-        return self.load_message_rx_tx(com_pdu_id_ref,
-                                       'CanIfTxPduCanId',
-                                       'CanIfTxPduDlc',
-                                       'CanIfTxPduCanIdType')
-
-    def load_message_rx(self, com_pdu_id_ref):
-        return self.load_message_rx_tx(com_pdu_id_ref,
-                                       'CanIfRxPduCanId',
-                                       'CanIfRxPduDlc',
-                                       'CanIfRxPduCanIdType')
-
-    def load_message_rx_tx(self,
-                           com_pdu_id_ref,
-                           parameter_can_id,
-                           parameter_dlc,
-                           parameter_can_id_type):
-        can_if_tx_pdu_cfg = self.find_can_if_rx_tx_pdu_cfg(com_pdu_id_ref)
-        frame_id = None
-        length = None
-        is_extended_frame = None
-
-        if can_if_tx_pdu_cfg is not None:
-            for parameter, value in self.iter_parameter_values(can_if_tx_pdu_cfg):
-                if parameter == parameter_can_id:
-                    frame_id = int(value)
-                elif parameter == parameter_dlc:
-                    length = int(value)
-                elif parameter == parameter_can_id_type:
-                    is_extended_frame = (value == 'EXTENDED_CAN')
-
-        return frame_id, length, is_extended_frame
-
-    def load_signal(self, xpath):
-        ecuc_container_value = self.find_value(xpath)
-        if ecuc_container_value is None:
-            return None
-
-        name = ecuc_container_value.find(SHORT_NAME_XPATH, NAMESPACES).text
-
-        # Default values.
-        is_signed = False
-        is_float = False
-        minimum = None
-        maximum = None
-        factor = 1.0
-        offset = 0.0
-        unit = None
-        choices = None
-        comments = None
-        receivers = []
-        decimal = SignalDecimal(Decimal(factor), Decimal(offset))
-
-        # Bit position, length, byte order, is_signed and is_float.
-        bit_position = None
-        length = None
-        byte_order = None
-
-        for parameter, value in self.iter_parameter_values(ecuc_container_value):
-            if parameter == 'ComBitPosition':
-                bit_position = int(value)
-            elif parameter == 'ComBitSize':
-                length = int(value)
-            elif parameter == 'ComSignalEndianness':
-                byte_order = value.lower()
-            elif parameter == 'ComSignalType':
-                if value in ['SINT8', 'SINT16', 'SINT32']:
-                    is_signed = True
-                elif value in ['FLOAT32', 'FLOAT64']:
-                    is_float = True
-
-        if bit_position is None:
-            LOGGER.warning('No bit position found for signal %s.',name)
-
-            return None
-
-        if length is None:
-            LOGGER.warning('No bit size found for signal %s.', name)
-
-            return None
-
-        if byte_order is None:
-            LOGGER.warning('No endianness found for signal %s.', name)
-
-            return None
-
-        # ToDo: minimum, maximum, factor, offset, unit, choices,
-        #       comments and receivers.
-
-        return Signal(name=name,
-                      start=bit_position,
-                      length=length,
-                      receivers=receivers,
-                      byte_order=byte_order,
-                      is_signed=is_signed,
-                      scale=factor,
-                      offset=offset,
-                      minimum=minimum,
-                      maximum=maximum,
-                      unit=unit,
-                      choices=choices,
-                      comment=comments,
-                      is_float=is_float,
-                      decimal=decimal)
-
-    def find_com_config(self, xpath):
-        return self.root.find(make_xpath([
-            "AR-PACKAGES",
-            "AR-PACKAGE/[ns:SHORT-NAME='{}']".format(xpath.split('/')[1]),
-            "ELEMENTS",
-            "ECUC-MODULE-CONFIGURATION-VALUES/[ns:SHORT-NAME='Com']",
-            "CONTAINERS",
-            "ECUC-CONTAINER-VALUE/[ns:SHORT-NAME='ComConfig']",
-            "SUB-CONTAINERS"
-        ]),
-                              NAMESPACES)
-
-    def find_value(self, xpath):
-        return self.root.find(make_xpath([
-            "AR-PACKAGES",
-            "AR-PACKAGE/[ns:SHORT-NAME='{}']".format(xpath.split('/')[1]),
-            "ELEMENTS",
-            "ECUC-MODULE-CONFIGURATION-VALUES/[ns:SHORT-NAME='Com']",
-            "CONTAINERS",
-            "ECUC-CONTAINER-VALUE/[ns:SHORT-NAME='ComConfig']",
-            "SUB-CONTAINERS",
-            "ECUC-CONTAINER-VALUE/[ns:SHORT-NAME='{}']".format(xpath.split('/')[-1])
-        ]),
-                              NAMESPACES)
-
-    def find_can_if_rx_tx_pdu_cfg(self, com_pdu_id_ref):
-        messages = self.root.iterfind(
-            make_xpath([
-                "AR-PACKAGES",
-                "AR-PACKAGE/[ns:SHORT-NAME='{}']".format(
-                    com_pdu_id_ref.split('/')[1]),
-                "ELEMENTS",
-                "ECUC-MODULE-CONFIGURATION-VALUES/[ns:SHORT-NAME='CanIf']",
-                'CONTAINERS',
-                "ECUC-CONTAINER-VALUE/[ns:SHORT-NAME='CanIfInitCfg']",
-                'SUB-CONTAINERS',
-                'ECUC-CONTAINER-VALUE'
-            ]),
-            NAMESPACES)
-
-        for message in messages:
-            definition_ref = message.find(DEFINITION_REF_XPATH,
-                                          NAMESPACES).text
-
-            if definition_ref.endswith('CanIfTxPduCfg'):
-                expected_reference = 'CanIfTxPduRef'
-            elif definition_ref.endswith('CanIfRxPduCfg'):
-                expected_reference = 'CanIfRxPduRef'
-            else:
-                continue
-
-            for reference, value in self.iter_reference_values(message):
-                if reference == expected_reference:
-                    if value == com_pdu_id_ref:
-                        return message
-
-    def iter_parameter_values(self, param_conf_container):
-        parameters = param_conf_container.find(PARAMETER_VALUES_XPATH,
-                                               NAMESPACES)
-
-        if parameters is None:
-            raise ValueError('PARAMETER-VALUES does not exist.')
-
-        for parameter in parameters:
-            definition_ref = parameter.find(DEFINITION_REF_XPATH,
-                                            NAMESPACES).text
-            value = parameter.find(VALUE_XPATH, NAMESPACES).text
-            name = definition_ref.split('/')[-1]
-
-            yield name, value
-
-    def iter_reference_values(self, param_conf_container):
-        references = param_conf_container.find(REFERENCE_VALUES_XPATH,
-                                               NAMESPACES)
-
-        if references is None:
-            raise ValueError('REFERENCE-VALUES does not exist.')
-
-        for reference in references:
-            definition_ref = reference.find(DEFINITION_REF_XPATH,
-                                            NAMESPACES).text
-            value = reference.find(VALUE_REF_XPATH, NAMESPACES).text
-            name = definition_ref.split('/')[-1]
-
-            yield name, value
-
-def is_ecu_extract(root: Any # For whatever reason, mypy does not
-                             # accept 'ElementTree' here...
-                   ) -> bool:
-    """Given the root object of an ARXML file's ElementTree,
-    determine if the file represents an ECU extract.
-
-    If it is not, it probably represents a system. Be aware that
-    currently loading ECU extracts is only supported for AUTOSAR 4.
-    """
-
-    ecuc_value_collection_xpath = \
-        './ns:AR-PACKAGES' + \
-        '/ns:AR-PACKAGE' + \
-        '/ns:ELEMENTS' + \
-        '/ns:ECUC-VALUE-COLLECTION'
-
-    namespaces = { 'ns': 'http://autosar.org/schema/r4.0' }
-
-    ecuc_value_collection = \
-        root.find(ecuc_value_collection_xpath, namespaces)
-
-    return ecuc_value_collection is not None
-
-def load_string(string:str,
-                strict:bool=True,
-                sort_signals:type_sort_signals=sort_signals_by_start_bit) \
-            -> InternalDatabase:
-    """Parse given ARXML format string.
-
-    """
-
-    root = ElementTree.fromstring(string)
-
-    m = re.match(r'{(.*)}AUTOSAR', root.tag)
-    if not m:
-        raise ValueError(f"No XML namespace specified or illegal root tag name '{root.tag}'")
-    xml_namespace = m.group(1)
-
-    # Should be replaced with a validation using the XSD file.
-    recognized_namespace = False
-    if re.match(r'http://autosar.org/schema/r(4.*)', xml_namespace) \
-       or re.match(r'http://autosar.org/(3.*)', xml_namespace) \
-       or re.match(r'http://autosar.org/(.*)\.DAI\.[0-9]', xml_namespace):
-        recognized_namespace = True
-
-    if not recognized_namespace:
-        raise ValueError(f"Unrecognized XML namespace '{xml_namespace}'")
-
-    if is_ecu_extract(root):
-        expected_root = f'{{{xml_namespace}}}AUTOSAR'
-        if root.tag != expected_root:
-            raise ValueError(f'Expected root element tag {expected_root}, '
-                             f'but got {root.tag}.')
-
-        return EcuExtractLoader(root, strict, sort_signals).load()
-    else:
-        return SystemLoader(root, strict, sort_signals).load()
